Bug fixes and UX work
Feedback page now shows alert of submission fail and success. If successful, feedback textarea clears so that there's no repeat submissions.

Search bar now shows alert when player is not found and tells user to be the first to contribute a rating for the player.

Changed contribute button and nav to "Contribute Rating" for better understanding.

Fixed contribute rating to only accept 0-10 rating reviews. Added a backend alert if some ratings are lower or higher.

Now showing total amount of ratings on home page.

Added social links at the bottom of home page.